simulation_claude_code_skill/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pyyaml
nltk
scikit-learn
python-multipart